requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
bcrypt==4.0.1
black==25.1.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.0
email_validator==2.2.0
executing==2.2.0
Faker==36.1.1
fastapi==0.115.8
flake8==7.1.2
fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.4.8
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
pywin32==308
pyzmq==26.2.1
rsa==4.9
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
win32_setctime==1.2.0
